[refactor] Move typing-related modules to typing package
