+++ src/model/actor.py
@@ -11,8 +11,8 @@
 from enum import Enum, auto
 from typing import Optional
 
-from src.model.dataclass_with_properties import dataclass_with_properties
-from src.model.type_checks import check_types_and_set_values
+from src.model.typing.dataclass_with_properties import dataclass_with_properties
+from src.model.typing.type_checks import check_types_and_set_values
 
 
 class ActorType(Enum):
