+++ src/model/extracted_licensing_info.py
@@ -11,8 +11,8 @@
 from dataclasses import field
 from typing import Optional, List
 
-from src.model.dataclass_with_properties import dataclass_with_properties
-from src.model.type_checks import check_types_and_set_values
+from src.model.typing.dataclass_with_properties import dataclass_with_properties
+from src.model.typing.type_checks import check_types_and_set_values
 
 
 @dataclass_with_properties
