+++ src/model/snippet.py
@@ -11,11 +11,11 @@
 from dataclasses import field
 from typing import Tuple, Optional, List, Union
 
-from src.model.dataclass_with_properties import dataclass_with_properties
+from src.model.typing.dataclass_with_properties import dataclass_with_properties
 from src.model.license_expression import LicenseExpression
 from src.model.spdx_no_assertion import SpdxNoAssertion
 from src.model.spdx_none import SpdxNone
-from src.model.type_checks import check_types_and_set_values
+from src.model.typing.type_checks import check_types_and_set_values
 
 
 @dataclass_with_properties
