+++ src/model/typing/dataclass_with_properties.py
